+++ projects/TrustTunnel/TrustTunnelClient/trusttunnel/setup_wizard/src/template_settings.rs
@@ -145,6 +145,10 @@ excluded_routes = [{}]
 mtu_size = {}
 {}
 change_system_dns = {}
+{}
+device_name = "{}"
+{}
+use_existing = {}
 "#,
         TunListener::doc_bound_if().to_toml_comment(),
         TunListener::default_bound_if(),
@@ -164,5 +168,9 @@ change_system_dns = {}
         TunListener::default_mtu_size(),
         TunListener::doc_change_system_dns().to_toml_comment(),
         TunListener::default_change_system_dns(),
+        TunListener::doc_device_name().to_toml_comment(),
+        TunListener::default_device_name(),
+        TunListener::doc_use_existing().to_toml_comment(),
+        TunListener::default_use_existing(),
     )
 });